book-recommendation-extractor-api/requirements.txt
numpy
Cython
fastapi==0.62.0
gunicorn
uvicorn
starlette==0.13.6
pydantic==1.8.2
spacy
spacy-transformers
aiofiles==0.6.0
python-multipart==0.0.5
jinja2==2.11.2
amightygirl.paapi5-python-sdk
python-dotenv==0.17.0